Support for filtering offers for longer periods.
Many offers will never be used, and can be filtered for
longer periods of time.  For example, a 1 CPU offer will
never match a broker with 2 CPUs.  We can also decline
offers for hosts that don't match a sticky constraint
for the duration left on the constraint.

In addition, in versions of mesos that support it (>0.25.0)
we can completely suppress offers when we know we'll never
do anything with them.

When the cluster is in "steady state" (eg all brokers are
in the state they should be in), we can tell mesos to stop
sending offers since we know we'll never need them.  Once
the scheduler detects it will need offers again, we notify
mesos via reviveOffers().